Feat: renamed basic pages to include the word template and implemented some minor visual tweaks
